+++ examples/nextjs-whiteboard-advanced/src/components/Ellipse.tsx
@@ -8,12 +8,9 @@ type Props = {
   selectionColor?: string;
 };
 
-export default function Ellipse({
-  layer,
-  onPointerDown,
-  id,
-  selectionColor,
-}: Props) {
+export default function Ellipse(
+  { layer, onPointerDown, id, selectionColor }: Props
+) {
   return (
     <ellipse
       onPointerDown={(e) => onPointerDown(e, id)}