+++ examples/nextjs-whiteboard-advanced/src/components/Rectangle.tsx
@@ -8,12 +8,9 @@ type Props = {
   selectionColor?: string;
 };
 
-export default function Rectangle({
-  layer,
-  onPointerDown,
-  id,
-  selectionColor,
-}: Props) {
+export default function Rectangle(
+  { layer, onPointerDown, id, selectionColor }: Props
+) {
   const { x, y, width, height, fill } = layer;
 
   return (