+++ examples/nextjs-whiteboard-advanced/src/components/SelectionTools.tsx
@@ -13,11 +13,9 @@ type SelectionToolsProps = {
   setLastUsedColor: (color: Color) => void;
 };
 
-function SelectionTools({
-  isAnimated,
-  camera,
-  setLastUsedColor,
-}: SelectionToolsProps) {
+function SelectionTools(
+  { isAnimated, camera, setLastUsedColor }: SelectionToolsProps
+) {
   const selection = useSelf((me) => me.presence.selection);
 
   /**