+++ packages/liveblocks-core/src/comments/types/BaseMetadata.ts
@@ -1,0 +1,1 @@
+export type BaseMetadata = Record<string, string | boolean | number>;