+++ packages/liveblocks-core/src/room.ts
@@ -1,4 +1,6 @@
 import type { AuthManager, AuthValue } from "./auth-manager";
+import type { CommentsApi } from "./comments";
+import { createCommentsApi } from "./comments";
 import type {
   Delegates,
   LegacyConnectionStatus,
@@ -28,14 +30,15 @@ import { isJsonArray, isJsonObject } from "./lib/Json";
 import { asPos } from "./lib/position";
 import type { Resolve } from "./lib/Resolve";
 import { compact, deepClone, tryParseJson } from "./lib/utils";
-import { canWriteStorage, TokenKind } from "./protocol/AuthToken";
-import type { BaseUserMeta } from "./protocol/BaseUserMeta";
+import { canComment, canWriteStorage, TokenKind } from "./protocol/AuthToken";
+import type { BaseUserInfo, BaseUserMeta } from "./protocol/BaseUserMeta";
 import type { ClientMsg } from "./protocol/ClientMsg";
 import { ClientMsgCode } from "./protocol/ClientMsg";
 import type { Op } from "./protocol/Op";
 import { isAckOp, OpCode } from "./protocol/Op";
 import type { IdTuple, SerializedCrdt } from "./protocol/SerializedCrdt";
 import type {
+  CommentsEventServerMsg,
   InitialDocumentStateServerMsg,
   RoomStateServerMsg,
   ServerMsg,
@@ -57,7 +60,7 @@ import type {
   IWebSocketMessageEvent,
 } from "./types/IWebSocket";
 import type { NodeMap } from "./types/NodeMap";
-import type { Others, OthersEvent } from "./types/Others";
+import type { OthersEvent } from "./types/Others";
 import type { User } from "./types/User";
 import { PKG_VERSION } from "./version";
 
@@ -93,7 +96,7 @@ type RoomEventCallbackMap<
   // since this API historically has taken _two_ callback arguments instead of
   // just one.
   others: (
-    others: Others<TPresence, TUserMeta>,
+    others: readonly User<TPresence, TUserMeta>[],
     event: OthersEvent<TPresence, TUserMeta>
   ) => void;
   error: Callback<Error>;
@@ -256,7 +259,7 @@ type SubscribeFn<
   (
     type: "others",
     listener: (
-      others: Others<TPresence, TUserMeta>,
+      others: readonly User<TPresence, TUserMeta>[],
       event: OthersEvent<TPresence, TUserMeta>
     ) => void
   ): () => void;
@@ -412,7 +415,7 @@ export type Room<
   TStorage extends LsonObject,
   TUserMeta extends BaseUserMeta,
   TRoomEvent extends Json,
-> = {
+> = CommentsApi<any /* TODO: Remove this any by adding a proper thread metadata on the Room type */> & {
   /**
    * @internal
    * Private methods to directly control the underlying state machine for this
@@ -480,7 +483,7 @@ export type Room<
    * @example
    * const others = room.getOthers();
    */
-  getOthers(): Others<TPresence, TUserMeta>;
+  getOthers(): readonly User<TPresence, TUserMeta>[];
 
   /**
    * Updates the presence of the current user. Only pass the properties you want to update. No need to send the full presence.
@@ -563,7 +566,7 @@ export type Room<
     readonly customEvent: Observable<{ connectionId: number; event: TRoomEvent; }>; // prettier-ignore
     readonly self: Observable<User<TPresence, TUserMeta>>;
     readonly myPresence: Observable<TPresence>;
-    readonly others: Observable<{ others: Others<TPresence, TUserMeta>; event: OthersEvent<TPresence, TUserMeta>; }>; // prettier-ignore
+    readonly others: Observable<{ others: readonly User<TPresence, TUserMeta>[]; event: OthersEvent<TPresence, TUserMeta>; }>; // prettier-ignore
     readonly error: Observable<Error>;
     readonly storage: Observable<StorageUpdate[]>;
     readonly history: Observable<HistoryEvent>;
@@ -577,6 +580,8 @@ export type Room<
 
     readonly storageStatus: Observable<StorageStatus>;
     readonly ydoc: Observable<YDocUpdate>;
+
+    readonly comments: Observable<CommentsEventServerMsg>;
   };
 
   /**
@@ -686,7 +691,7 @@ type IdFactory = () => string;
 
 type StaticSessionInfo = {
   readonly userId?: string;
-  readonly userInfo?: Json;
+  readonly userInfo?: BaseUserInfo;
 };
 
 type DynamicSessionInfo = {
@@ -1112,7 +1117,7 @@ export function createRoom<
     self: makeEventSource<User<TPresence, TUserMeta>>(),
     myPresence: makeEventSource<TPresence>(),
     others: makeEventSource<{
-      others: Others<TPresence, TUserMeta>;
+      others: readonly User<TPresence, TUserMeta>[];
       event: OthersEvent<TPresence, TUserMeta>;
     }>(),
     error: makeEventSource<Error>(),
@@ -1121,6 +1126,8 @@ export function createRoom<
     storageDidLoad: makeEventSource<void>(),
     storageStatus: makeEventSource<StorageStatus>(),
     ydoc: makeEventSource<YDocUpdate>(),
+
+    comments: makeEventSource<CommentsEventServerMsg>(),
   };
 
   function sendMessages(
@@ -1177,6 +1184,7 @@ export function createRoom<
           info: staticSession.userInfo,
           presence: myPresence,
           canWrite,
+          canComment: canComment(dynamicSession.scopes),
           isReadOnly: !canWrite, // Deprecated, kept for backward-compatibility
         };
       }
@@ -1814,6 +1822,15 @@ export function createRoom<
 
             break;
           }
+
+          case ServerMsgCode.THREAD_CREATED:
+          case ServerMsgCode.THREAD_METADATA_UPDATED:
+          case ServerMsgCode.COMMENT_CREATED:
+          case ServerMsgCode.COMMENT_EDITED:
+          case ServerMsgCode.COMMENT_DELETED: {
+            eventHub.comments.notify(message);
+            break;
+          }
         }
       }
 
@@ -2173,7 +2190,13 @@ export function createRoom<
     storageDidLoad: eventHub.storageDidLoad.observable,
     storageStatus: eventHub.storageStatus.observable,
     ydoc: eventHub.ydoc.observable,
+
+    comments: eventHub.comments.observable,
   };
+
+  const commentsApi = createCommentsApi(config.roomId, delegates.authenticate, {
+    serverEndpoint: "https://api.liveblocks.io/v2",
+  });
 
   return Object.defineProperty(
     {
@@ -2235,6 +2258,8 @@ export function createRoom<
       // Presence
       getPresence: () => context.myPresence.current,
       getOthers: () => context.others.current,
+
+      ...commentsApi,
     },
 
     // Explictly make the __internal field non-enumerable, to avoid aggressive
@@ -2255,7 +2280,10 @@ function makeClassicSubscribeFn<
   TUserMeta extends BaseUserMeta,
   TRoomEvent extends Json,
 >(
-  events: Room<TPresence, TStorage, TUserMeta, TRoomEvent>["events"]
+  events: Omit<
+    Room<TPresence, TStorage, TUserMeta, TRoomEvent>["events"],
+    "comments" // comments is an internal events so we omit it from the subscribe method
+  >
 ): SubscribeFn<TPresence, TStorage, TUserMeta, TRoomEvent> {
   // Set up the "subscribe" wrapper API
   function subscribeToLiveStructureDeeply<L extends LiveStructure>(
@@ -2316,7 +2344,7 @@ function makeClassicSubscribeFn<
           // NOTE: Others have a different callback structure, where the API
           // exposed on the outside takes _two_ callback arguments!
           const cb = callback as (
-            others: Others<TPresence, TUserMeta>,
+            others: readonly User<TPresence, TUserMeta>[],
             event: OthersEvent<TPresence, TUserMeta>
           ) => void;
           return events.others.subscribe(({ others, event }) =>