+++ packages/liveblocks-node/src/index.ts
@@ -1,8 +1,15 @@
 export { authorize } from "./authorize";
-export type { LiveblocksOptions } from "./new-auth";
-export { Liveblocks } from "./new-auth";
+export type { LiveblocksOptions } from "./client";
+export { Liveblocks } from "./client";
 export type {
+  CommentCreatedEvent,
+  CommentDeletedEvent,
+  CommentEditedEvent,
+  RoomCreatedEvent,
+  RoomDeletedEvent,
   StorageUpdatedEvent,
+  ThreadCreatedEvent,
+  ThreadMetadataUpdatedEvent,
   UserEnteredEvent,
   UserLeftEvent,
   WebhookEvent,