+++ packages/liveblocks-react-comments/src/utils/wrap.ts
@@ -1,0 +1,5 @@
+export function wrap(value: number, min: number, max: number) {
+  const range = max - min;
+
+  return range > 0 ? ((((value - min) % range) + range) % range) + min : 0;
+}