+++ packages/liveblocks-zustand/src/__tests__/index.test.ts
@@ -351,6 +351,7 @@ describe("middleware", () => {
             name: "Testy McTester",
           },
           canWrite: true,
+          canComment: true,
           isReadOnly: false,
           presence: { x: 1 },
         },